requirements.txt
pyjwt
paho-mqtt